Refactor MRZ validation logic and add new validation utility; update version to 0.0.2+4
