+++ .flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"camerawesome","path":"/Users/juanvegu/.pub-cache/hosted/pub.dev/camerawesome-2.1.0/","native_build":true,"dependencies":[]},{"name":"google_mlkit_commons","path":"/Users/juanvegu/.pub-cache/hosted/pub.dev/google_mlkit_commons-0.9.0/","native_build":true,"dependencies":[]},{"name":"google_mlkit_text_recognition","path":"/Users/juanvegu/.pub-cache/hosted/pub.dev/google_mlkit_text_recognition-0.14.0/","native_build":true,"dependencies":["google_mlkit_commons"]},{"name":"path_provider_foundation","path":"/Users/juanvegu/.pub-cache/hosted/pub.dev/path_provider_foundation-2.4.0/","shared_darwin_source":true,"native_build":true,"dependencies":[]}],"android":[{"name":"camerawesome","path":"/Users/juanvegu/.pub-cache/hosted/pub.dev/camerawesome-2.1.0/","native_build":true,"dependencies":[]},{"name":"google_mlkit_commons","path":"/Users/juanvegu/.pub-cache/hosted/pub.dev/google_mlkit_commons-0.9.0/","native_build":true,"dependencies":[]},{"name":"google_mlkit_text_recognition","path":"/Users/juanvegu/.pub-cache/hosted/pub.dev/google_mlkit_text_recognition-0.14.0/","native_build":true,"dependencies":["google_mlkit_commons"]},{"name":"path_provider_android","path":"/Users/juanvegu/.pub-cache/hosted/pub.dev/path_provider_android-2.2.14/","native_build":true,"dependencies":[]}],"macos":[{"name":"path_provider_foundation","path":"/Users/juanvegu/.pub-cache/hosted/pub.dev/path_provider_foundation-2.4.0/","shared_darwin_source":true,"native_build":true,"dependencies":[]}],"linux":[{"name":"path_provider_linux","path":"/Users/juanvegu/.pub-cache/hosted/pub.dev/path_provider_linux-2.2.1/","native_build":false,"dependencies":[]}],"windows":[{"name":"path_provider_windows","path":"/Users/juanvegu/.pub-cache/hosted/pub.dev/path_provider_windows-2.3.0/","native_build":false,"dependencies":[]}],"web":[]},"dependencyGraph":[{"name":"camerawesome","dependencies":["path_provider"]},{"name":"google_mlkit_commons","dependencies":[]},{"name":"google_mlkit_text_recognition","dependencies":["google_mlkit_commons"]},{"name":"path_provider","dependencies":["path_provider_android","path_provider_foundation","path_provider_linux","path_provider_windows"]},{"name":"path_provider_android","dependencies":[]},{"name":"path_provider_foundation","dependencies":[]},{"name":"path_provider_linux","dependencies":[]},{"name":"path_provider_windows","dependencies":[]}],"date_created":"2024-11-25 21:18:16.562496","version":"3.24.5","swift_package_manager_enabled":false}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"camerawesome","path":"/Users/juanvegu/.pub-cache/hosted/pub.dev/camerawesome-2.1.0/","native_build":true,"dependencies":[]},{"name":"google_mlkit_commons","path":"/Users/juanvegu/.pub-cache/hosted/pub.dev/google_mlkit_commons-0.9.0/","native_build":true,"dependencies":[]},{"name":"google_mlkit_text_recognition","path":"/Users/juanvegu/.pub-cache/hosted/pub.dev/google_mlkit_text_recognition-0.14.0/","native_build":true,"dependencies":["google_mlkit_commons"]},{"name":"path_provider_foundation","path":"/Users/juanvegu/.pub-cache/hosted/pub.dev/path_provider_foundation-2.4.0/","shared_darwin_source":true,"native_build":true,"dependencies":[]}],"android":[{"name":"camerawesome","path":"/Users/juanvegu/.pub-cache/hosted/pub.dev/camerawesome-2.1.0/","native_build":true,"dependencies":[]},{"name":"google_mlkit_commons","path":"/Users/juanvegu/.pub-cache/hosted/pub.dev/google_mlkit_commons-0.9.0/","native_build":true,"dependencies":[]},{"name":"google_mlkit_text_recognition","path":"/Users/juanvegu/.pub-cache/hosted/pub.dev/google_mlkit_text_recognition-0.14.0/","native_build":true,"dependencies":["google_mlkit_commons"]},{"name":"path_provider_android","path":"/Users/juanvegu/.pub-cache/hosted/pub.dev/path_provider_android-2.2.14/","native_build":true,"dependencies":[]}],"macos":[{"name":"path_provider_foundation","path":"/Users/juanvegu/.pub-cache/hosted/pub.dev/path_provider_foundation-2.4.0/","shared_darwin_source":true,"native_build":true,"dependencies":[]}],"linux":[{"name":"path_provider_linux","path":"/Users/juanvegu/.pub-cache/hosted/pub.dev/path_provider_linux-2.2.1/","native_build":false,"dependencies":[]}],"windows":[{"name":"path_provider_windows","path":"/Users/juanvegu/.pub-cache/hosted/pub.dev/path_provider_windows-2.3.0/","native_build":false,"dependencies":[]}],"web":[]},"dependencyGraph":[{"name":"camerawesome","dependencies":["path_provider"]},{"name":"google_mlkit_commons","dependencies":[]},{"name":"google_mlkit_text_recognition","dependencies":["google_mlkit_commons"]},{"name":"path_provider","dependencies":["path_provider_android","path_provider_foundation","path_provider_linux","path_provider_windows"]},{"name":"path_provider_android","dependencies":[]},{"name":"path_provider_foundation","dependencies":[]},{"name":"path_provider_linux","dependencies":[]},{"name":"path_provider_windows","dependencies":[]}],"date_created":"2024-11-30 15:06:09.615066","version":"3.24.5","swift_package_manager_enabled":false}
