+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: mrz_scanner
 description: A package that allows you to Scan any kind of documents that have any mrz format
-version: 0.0.2+3
+version: 0.0.2+4
 homepage:
 
 environment:
